micro_profiling/latency/requirements.txt
google-cloud-logging
boto3
s3-log-parse
azure-monitor-opentelemetry
azure-monitor-query
azure-identity